Pos (#20)
* fix the terminal payment

* fix all payments

* adding relocate

* update products module

* updated commit pos module

* fix the pos payment

* cart payment updated

* fix image

* fix the products & pos

* fix station registers

* fixing station module

* updated pos

* fix merge

* real-time cart deletions

* send email receipt

* fix image products

* reservation multi-step fix
gitinigore storage

* fix

* Delete storage/app/public/products directory

---------

+++ database/migrations/2020_04_30_053758_create_user_cart_table.php
@@ -13,15 +13,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('admin_cart', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('admin_id');
-            $table->foreignId('product_id');
-            $table->unsignedInteger('quantity');
+        // Check if the table already exists to avoid trying to recreate it
+        if (!Schema::hasTable('admin_cart')) {
+            Schema::create('admin_cart', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('admin_id');
+                $table->foreignId('product_id');
+                $table->unsignedInteger('quantity');
 
-            $table->foreign('admin_id')->references('id')->on('admins')->onDelete('cascade');
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-        });
+                $table->foreign('admin_id')->references('id')->on('admins')->onDelete('cascade');
+                $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
+            });
+        }
     }
 
     /**
